requirements.txt
Flask==2.0.2
Pillow==8.2.0
